move file pkg/hints/usort_hint_codes.go -> pkg/hints/hint_codes/usort_hint_codes.go
